Current User stuff done

+++ SOS/src/main/java/com/Ajwain/SOS/services/DeadlineService.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.Ajwain.SOS.auth.CurrentUserService;
 import com.Ajwain.SOS.config.PaginationConfig;
 import com.Ajwain.SOS.dto.DeadlineRequestDTO;
 import com.Ajwain.SOS.dto.DeadlineResponseDTO;
@@ -19,6 +20,7 @@ import com.Ajwain.SOS.dto.DeadlineSearchCriteria;
 import com.Ajwain.SOS.dto.PaginationResponseDTO;
 import com.Ajwain.SOS.entities.Deadline;
 import com.Ajwain.SOS.entities.Subject;
+import com.Ajwain.SOS.entities.User;
 import com.Ajwain.SOS.entities.enums.DeadlineType;
 import com.Ajwain.SOS.exception.BadRequestException;
 import com.Ajwain.SOS.exception.ResourceNotFoundException;
@@ -32,17 +34,26 @@ import jakarta.transaction.Transactional;
 public class DeadlineService {
 	private final SubjectRepository subjectRepository;
 	private final DeadlineRepository deadlineRepository;
+	private final CurrentUserService currentUserService;
 	private final Logger logger=LoggerFactory.getLogger(StudyPlanService.class);
 	
 	private final StudyPlanService studyPlanService;
-	public DeadlineService(SubjectRepository subjectRepository,DeadlineRepository deadlineRepository,StudyPlanService studyPlanService) {
+	public DeadlineService(SubjectRepository subjectRepository,DeadlineRepository deadlineRepository,StudyPlanService studyPlanService ,CurrentUserService currentUserService) {
 		this.subjectRepository=subjectRepository;
+		this.currentUserService=currentUserService;
 		this.deadlineRepository=deadlineRepository;
 		this.studyPlanService=studyPlanService;
 	}
 	public DeadlineResponseDTO createDeadline(long subjectId,DeadlineRequestDTO dto) {
 		Deadline deadline =new Deadline();
-		Subject subject=subjectRepository.findById(subjectId).orElseThrow(()->new ResourceNotFoundException("Subject not found"));
+		User user = currentUserService.getCurrentUser();
+
+		Subject subject = subjectRepository.findById(subjectId)
+		    .orElseThrow(() -> new ResourceNotFoundException("Subject not found"));
+
+		if (subject.getUser().getId()!=(user.getId())) {
+		    throw new BadRequestException("Unauthorized access to subject");
+		}
 		if(dto.getDeadlineDate().isBefore(LocalDateTime.now())) {
 		    throw new BadRequestException("Deadline cannot be in the past");
 		}
@@ -52,24 +63,39 @@ public class DeadlineService {
 		deadline.setDeadlineType(dto.getDeadlineType());
 		deadline.setSubject(subject);
 		Deadline savedDeadline=deadlineRepository.save(deadline);
-		studyPlanService.regenerateStudyPlan(deadline.getSubject().getUser().getId());
-
+		studyPlanService.regenerateStudyPlan(user.getId());
 		logger.info("Creating deadline for subject {}", subjectId);
 		return convertToResponseDTO(savedDeadline);
 	}
-	public List<DeadlineResponseDTO> getDeadlinesByUser(long userId){
-		
-		return deadlineRepository.findBySubjectUserId(userId).stream().map(this::convertToResponseDTO).toList();
+	public List<DeadlineResponseDTO> getDeadlinesByUser(){
+	    User user = currentUserService.getCurrentUser();
+	    return deadlineRepository.findBySubjectUser(user)
+	            .stream().map(this::convertToResponseDTO).toList();
 	}
 	public List<DeadlineResponseDTO> getDeadlinesBySubject(long subjectId){
-		subjectRepository.findById(subjectId)
-	    .orElseThrow(() -> new ResourceNotFoundException("Subject not found"));
-		return deadlineRepository.findBySubjectId(subjectId).stream().map(this::convertToResponseDTO).toList();
+	    User user = currentUserService.getCurrentUser();
+
+	    Subject subject = subjectRepository.findById(subjectId)
+	        .orElseThrow(() -> new ResourceNotFoundException("Subject not found"));
+
+	    if (subject.getUser().getId()!=(user.getId())) {
+	        throw new BadRequestException("Unauthorized access");
+	    }
+
+	    return deadlineRepository.findBySubject(user, subject)
+	            .stream()
+	            .map(this::convertToResponseDTO)
+	            .toList();
 	}
 	public DeadlineResponseDTO updateDeadline(long deadlineID,DeadlineRequestDTO dto) {
 		Deadline deadline=deadlineRepository.findById(deadlineID).orElseThrow(()->new ResourceNotFoundException("Deadline not found"));
 		if(dto.getDeadlineDate().isBefore(LocalDateTime.now())) {
 		    throw new BadRequestException( "Deadline cannot be in the past");
+		}
+		User user = currentUserService.getCurrentUser();
+
+		if (deadline.getSubject().getUser().getId()!=(user.getId())) {
+		    throw new BadRequestException("Unauthorized");
 		}
 		deadline.setDeadlineDate(dto.getDeadlineDate());
 		deadline.setDeadlinePriority(dto.getDeadlinePriority());
@@ -77,7 +103,7 @@ public class DeadlineService {
 		deadline.setDeadlineType(dto.getDeadlineType());
 		
 		Deadline savedDeadline=deadlineRepository.save(deadline);
-		studyPlanService.regenerateStudyPlan(deadline.getSubject().getUser().getId());
+		studyPlanService.regenerateStudyPlan(user.getId());
 		logger.info("Updating deadline {}", deadlineID);
 		return convertToResponseDTO(savedDeadline);
 	}
@@ -85,16 +111,20 @@ public class DeadlineService {
 	public void deleteDeadline(long deadlineId) {
 	    Deadline deadline = deadlineRepository.findById(deadlineId)
 	        .orElseThrow(() -> new ResourceNotFoundException("Deadline not found"));
+	    User user = currentUserService.getCurrentUser();
 
+	    if (deadline.getSubject().getUser().getId()!=(user.getId())) {
+	        throw new BadRequestException("Unauthorized");
+	    }
 	    deadlineRepository.delete(deadline);
-	    studyPlanService.regenerateStudyPlan(deadline.getSubject().getUser().getId());
-
+	    studyPlanService.regenerateStudyPlan(user.getId());
 	    logger.info("Deleting deadline {}", deadlineId);
 	}
-	public List<DeadlineResponseDTO> getUpcomingDeadlines(long userId) {
+	public List<DeadlineResponseDTO> getUpcomingDeadlines() {
+	    User user = currentUserService.getCurrentUser();
 
 	    return deadlineRepository
-	        .findBySubjectUserIdAndDeadlineDateAfterOrderByDeadlineDateAsc(userId, LocalDateTime.now())
+	        .findBySubjectUserAndDeadlineDateAfterOrderByDeadlineDateAsc(user, LocalDateTime.now())
 	        .stream()
 	        .map(this::convertToResponseDTO)
 	        .toList();
@@ -108,11 +138,12 @@ public class DeadlineService {
 		List<DeadlineResponseDTO> dtos=deadline.getContent().stream().map(this::convertToResponseDTO).toList();
 		return PaginationResponseDTO.fromPage(deadline,dtos);
 	}
-	public PaginationResponseDTO<DeadlineResponseDTO> getDeadlinesByUser(long userId,Pageable pageable){
+	public PaginationResponseDTO<DeadlineResponseDTO> getDeadlinesByUser(Pageable pageable){
 		pageable=validatePageable(pageable);
-			
-	
-		Page<Deadline> deadline=deadlineRepository.findBySubjectUserId(userId,pageable);
+
+		User user = currentUserService.getCurrentUser();
+
+		Page<Deadline> deadline = deadlineRepository.findBySubjectUser(user, pageable);
 		List<DeadlineResponseDTO> dtos=deadline.getContent().stream().map(this::convertToResponseDTO).toList();
 		return PaginationResponseDTO.fromPage(deadline,dtos);
 	}
@@ -138,7 +169,8 @@ public class DeadlineService {
 	public PaginationResponseDTO<DeadlineResponseDTO> getDeadlines(DeadlineSearchCriteria criteria,
 	        Pageable pageable){
 		pageable=validatePageable(pageable);
-		Specification<Deadline> s=buildSpecification(criteria);
+		User user=currentUserService.getCurrentUser();
+		Specification<Deadline> s=buildSpecification(criteria).and(DeadlineSpecification.belongsToUser(user.getId()));
 		Page<Deadline> deadline=deadlineRepository.findAll(s,pageable);
 		List<DeadlineResponseDTO> dtos=deadline.getContent().stream().map(this::convertToResponseDTO).toList();
 		return PaginationResponseDTO.fromPage(deadline, dtos);
